Alter Ingredient table to tear apart htmlfrag
Turns out htmlfrag contained a bunch of boilerplate HTML we could get rid
of which helps to slim it down.  It also contained a category embedded in
a header which we extract onto the ingredient as a first class field.
Finally we also convert all the `x-cocktaildb://ingredients/<blah>` urls
in the descriptions into URLs for our structure.  It's annoying to have
url paths embedded in the content, but it's ok for now, and we can always
fix it later as I suspect there'll be something similar to do for the
recipe step descriptions.

Having done this we can update the ingredient page to use our new category
and description fields, and the browsing of ingredients now works!