Extract ingredient references from `xmlbody` and make models
The `xmlbody` of a recipe formulation contains an html-ish structure that
describes the steps to take for how to make the recipe.  In these steps
are xml `ing` tags that contain ids that point to either an `Ingredient`
or a `Recipe`.  It's not clear why this is the structure used here instead
of the `href="x-cocktaildb://ingredients/<blah>"` structure found in the
`htmlfrag` of the ingredients.  No matter though, we extract all the `ing`
tags for each recipe formulation and use them to populate a new table that
links recipe formulations to ingredients or other recipe formulations.

We can add that to the page for a recipe and now we can find out more
about ingredients on the recipe pages.