Convert `<meas>` tags into HTML measurement references
As these tags contain data for lots of different units, we render them all
and then provide a units UI for selecting which one to show.  By default
we choose ml because that makes most sense.  We use some JS to check what
is selected in the UI and then update what is shown and hidden
accordingly in the text.