Extract `<bwr>` references as Barware and link to them
The source db didn't include any obvious barware table or data so we have
to extract it all from the `xmlbody` on the recipe formulations.  All we
have right now are the name and some synonyms (which usually is the
pluralised version tbh so we might get rid of this later).