Introduce `BarwareImage`
This is backed by the `Image` table.  Originally I hoped that this would
be the data for the `bwr` tags in the `xmlbody` of the recipe formulations
but unfortunately it's not.  The images are silhouettes of barware, but
the `id` in the `Image` table is not the same as the `id` referenced in
the `bwr` tag.  It's actually the `id` of an entry in the cocktailsdb.com
website for barware.  Or, at least, it _was_ for the version of that
website in ~2009 when this app & db was created.  There's probably a link
somewhere between the `bwr` and `Image` - perhaps just in the barware item
name and the `image_name`.