By default, remove the ingredient characteristics
We've extracted explicit formulation -> ingredient links so we don't need
the characteristic versions.  We want the other characteristics though so
we add an activerecord enum for category and then define `without_xxx`
scope for each category value to exclude that characteristic category.
Finally, we make `without_ingredients` part of the default scope on the
`characteristics` association on recipes and formulations.