Provide link between characteristic and ingredient
Of interest to base characteristics only really, and we've fiddled the
data so that the base labels and their relevant ingredients match up
better.  This means we can link from the base page to the ingredient page,
or, later we could even pull in some of the details from the ingredient
to this page.