Add basic cocktails search
Powered by the [middleman-search-gds][] gem which provides static site
searching by building a [lunr.js][] index as /search.json.  We then
provide a client side JS search using lunr.js, our implementation of which
is _heavily_ inspired by the [search functionality added][gds-search-impl]
in the GDS [tech-docs-gem][] except we de-jqueried it and slightly
simplified it for our usecase (no ga tracking, etc...).

Just to get it working we index the content of the page, we'll finesse
this in future commits as the page content is imperfect.

[middleman-search-gds]: https://github.com/alphagov/middleman-search
[lunr.js]: https://lunrjs.com
[tech-docs-gem]: https://github.com/alphagov/tech-docs-gem/
[gds-search-impl]: https://github.com/alphagov/tech-docs-gem/blob/d077820ab9dec2373fea4761c9f52d0e76cfc46a/lib/assets/javascripts/_modules/search.js