Add a middleman extension to add ingredients metadata to recipe pages
The most useful version of the search we added will be to search by the
ingredients you have on hand (Gin, Campari, Vermouth), to see which
recipes you can make.  It's also useful to search by the title for those
cases where you know what you want to drink (a Negroni, obvs) but can't
remember how to make one.  To achieve this we need the ingredient details
and recipe title to be added as metadata to the pages we want to search
as that's the only way we can generate the index with this gem.  To do
this we introduce an extension that pre-processes any `recipes/` pages to
add metadata that includes a list of ingredients and the title of the
`Recipe` instance that backs the page.

I suspect there's a smarter way.  Maybe we shouldn't have proxy objects at
all but just use our new extension to create the pages for recipes.  Or
maybe we want a version of middleman-search that isn't tied to page
objects.